+++ src/components/BsField/BsTextField.vue
@@ -161,7 +161,7 @@ export default {
                 'md-floating-label': this.floatingLabel,
                 'md-focused': this.isFocused,
                 'has-error': this.hasValidationError,
-                'has-success': this.wasValidated && !this.hasValidationError
+                'has-success': this.hasValidated && !this.hasValidationError
             }
         },
         /**
@@ -212,6 +212,7 @@ export default {
     },
     watch: {
         value(newVal) {
+            this.localValue = newVal;
             this._updateLegend(newVal);
             // this._setFloatingLabelPosition();
         }